+++ YarnSpinner.Tests/ErrorHandlingTests.cs
@@ -79,5 +79,20 @@ namespace YarnSpinner.Tests
 
             result.Diagnostics.Should().Contain(d => d.Message.Contains(@"Unexpected "">>"" while reading a function call"));
         }
+
+        // testing that warnings are generated for empty nodes
+        [Fact]
+        public void TestEmptyNodesGenerateWarnings()
+        {
+            var source = CreateTestNode("", "Start");
+
+            var result = Compiler.Compile(CompilationJob.CreateFromString("<inputs>", source));
+
+            var warnings = result.Diagnostics.Where(d => d.Severity == Diagnostic.DiagnosticSeverity.Warning);
+            // there should be only one warning
+            warnings.Count().Should().Be(1);
+            // and it should be the warning about empty nodes
+            warnings.FirstOrDefault().Message.Should().Be("Node \"Start\" is empty and will not be included in the compiled output.");
+        }
     }
 }