+++ YarnSpinner.Tests/TypeTests.cs
@@ -996,8 +996,23 @@ namespace YarnSpinner.Tests
             resultConstraint.Children.ElementAt(1).Children.Should().BeEquivalentTo(new[] { a, d, e, f });
             resultConstraint.Children.ElementAt(2).Children.Should().BeEquivalentTo(new[] { b, c, f });
             resultConstraint.Children.ElementAt(3).Children.Should().BeEquivalentTo(new[] { b, d, e, f });
-
+        }
             
+        [Fact]
+        public void TestSelfReferencingSetErrors()
+        {
+            // this test should fail on the v3 type system which is good, but for now I want to make sure its tested as is
+            // plus when this fails on v3 it's a good reminder to remove the hack
+            var source = CreateTestNode("<<set $badInference = $badInference + 1>>");
+
+            var result = Compiler.Compile(CompilationJob.CreateFromString("input", source));
+
+            // we should have one diagnostic
+            result.Diagnostics.Count().Should().Be(1);
+            // it's an error
+            result.Diagnostics.FirstOrDefault().Severity.Should().Be(Diagnostic.DiagnosticSeverity.Error);
+            // and it's the "I found this twice so don't know what to do" error
+            result.Diagnostics.Should().Contain(d => d.Message.Contains("\"$badInference\" has had its default value inferred in multiple places:"));
         }
     }
 }